[Android] prepare single-only build
With the request to build single-shot only library, add definition and  update build script.
