+++ api/android/api/src/main/jni/Android.mk
@@ -45,14 +45,20 @@ include $(CLEAR_VARS)
 
 LOCAL_MODULE := nnstreamer-native
 LOCAL_SRC_FILES := nnstreamer-native-api.c \
-    nnstreamer-native-customfilter.c \
-    nnstreamer-native-pipeline.c \
     nnstreamer-native-singleshot.c
 LOCAL_CFLAGS += -O2 -DVERSION=\"$(NNSTREAMER_VERSION)\"
 LOCAL_C_INCLUDES := $(NNSTREAMER_INCLUDES) $(NNSTREAMER_CAPI_INCLUDES)
 LOCAL_STATIC_LIBRARIES := nnstreamer tensorflow-lite cpufeatures
 LOCAL_SHARED_LIBRARIES := gstreamer_android
 LOCAL_LDLIBS := -llog -landroid -lmediandk
+
+ifeq ($(NNSTREAMER_API_OPTION),single)
+LOCAL_CFLAGS += -DNNS_SINGLE_ONLY=1
+else
+LOCAL_SRC_FILES += \
+    nnstreamer-native-customfilter.c \
+    nnstreamer-native-pipeline.c
+endif
 
 ifeq ($(ENABLE_SNAP), true)
 LOCAL_CFLAGS += -DENABLE_SNAP=1
@@ -68,8 +74,12 @@ GSTREAMER_NDK_BUILD_PATH := $(GSTREAMER_ROOT)/share/gst-android/ndk-build/
 include $(LOCAL_PATH)/Android-gst-plugins.mk
 
 GSTREAMER_PLUGINS        := $(GST_REQUIRED_PLUGINS)
-GSTREAMER_EXTRA_DEPS     := $(GST_REQUIRED_DEPS)
-GSTREAMER_EXTRA_LIBS     := $(GST_REQUIRED_LIBS)
+GSTREAMER_EXTRA_DEPS     := $(GST_REQUIRED_DEPS) gio-2.0
+GSTREAMER_EXTRA_LIBS     := $(GST_REQUIRED_LIBS) -liconv
+
+GSTREAMER_INCLUDE_FONTS := no
+GSTREAMER_INCLUDE_CA_CERTIFICATES := no
+
 include $(GSTREAMER_NDK_BUILD_PATH)/gstreamer-1.0.mk
 
 $(call import-module, android/cpufeatures)
